requirements.txt
python-telegram-bot==21.0.1
playwright==1.49.0
requests==2.31.0
greenlet>=3.1.1
asyncio
